feat(feed): formatted description in dialog user card
